Created Apollo 14 menu item
